cleanup bao-video to use the new API
also condense main into a single routine again, now that
we have gone through the trouble of creating interoperable
APIs.
