+++ services/bao-video/src/api.rs
@@ -3,3 +3,5 @@ pub enum Opcode {
     CamIrq,
     InvalidCall,
 }
+
+pub const SERVER_NAME_GFX: &str = "_Graphics_";
